Separate files for nexo entities
