+++ custom_components/nexo/nexo_gate.py
@@ -1,7 +1,6 @@
-from nexoResource import NexoResource
+from nexo_resource import NexoResource
 
 
 class NexoGate(NexoResource):
     def __init__(self, web_socket, *args, **kwargs) -> None:
         super().__init__(web_socket, *args, **kwargs)
-
