+++ custom_components/nexo/nexo_group_dimmer.py
@@ -1,6 +1,7 @@
-from nexoResource import NexoResource
+from nexo_resource import NexoResource
 
-class NexoBlindGroup(NexoResource):
+
+class NexoGroupDimmer(NexoResource):
     def __init__(self, web_socket, ios, *args, **kwargs):
         super().__init__(web_socket, *args, **kwargs)
         self.ios = ios
